# Use zero-based numeric indices for much faster instrumented code