--- a/src/visitor.js
+++ b/src/visitor.js
@@ -91,7 +91,7 @@ class VisitState {
             return;
         }
         // else check custom node attribute set by a prior visitor
-        if (this.getAttr(path.node, 'skip-all')) {
+        if (this.getAttr(path.node, 'skip-all') !== null) {
             this.nextIgnore = n;
         }
     }
@@ -354,7 +354,7 @@ function coverSwitchCase(path) {
     const T = this.types;
     const b = this.getAttr(path.parentPath.node, 'branchName');
     /* istanbul ignore if: paranoid check */
-    if (!b) {
+    if (b === null) {
         throw new Error('Unable to get switch branch name');
     }
     const increment = this.getBranchIncrement(b, path.node.loc);