--- a/test/specs/arrow-fn.yaml
+++ b/test/specs/arrow-fn.yaml
@@ -9,15 +9,15 @@ tests:
     args: [1, 2, 3, 4]
     out: [1, 4, 9, 16]
     lines: {'1': 1, '2': 4}
-    statements: {'1': 1, '2': 1, '3': 4}
-    functions: {'1': 4}
+    statements: {'0': 1, '1': 1, '2': 4}
+    functions: {'0': 4}
 
   - name: function not called
     args: []
     out: []
     lines: {'1': 1, '2': 1}
-    statements: {'1': 1, '2': 1, '3': 0}
-    functions: {'1': 0}
+    statements: {'0': 1, '1': 1, '2': 0}
+    functions: {'0': 0}
 
 ---
 name:  es6 arrow function block
@@ -30,15 +30,15 @@ tests:
     args: [1, 2, 3, 4]
     out: [1, 4, 9, 16]
     lines: {'1': 1, '2': 4}
-    statements: {'1': 1, '2': 1, '3': 4}
-    functions: {'1': 4}
+    statements: {'0': 1, '1': 1, '2': 4}
+    functions: {'0': 4}
 
   - name: function not called
     args: []
     out: []
     lines: {'1': 1, '2': 1}
-    statements: {'1': 1, '2': 1, '3': 0}
-    functions: {'1': 0}
+    statements: {'0': 1, '1': 1, '2': 0}
+    functions: {'0': 0}
 
 ---
 name: complex arrow fn
@@ -57,6 +57,6 @@ tests:
     args: [1, 2, 3, 4]
     out: 10
     lines: { '1': 1, '2': 1, '3': 1, '4': 1, '8': 1}
-    functions: { '1': 1, '2': 1, '3': 1, '4': 1 }
-    statements: { '1': 1, '2': 1, '3': 1, '4': 1, '5': 1, '6':1 }
+    functions: { '0': 1, '1': 1, '2': 1, '3': 1 }
+    statements: { '0': 1, '1': 1, '2': 1, '3': 1, '4': 1, '5':1 }