--- a/test/specs/default-args.yaml
+++ b/test/specs/default-args.yaml
@@ -11,22 +11,22 @@ tests:
     args: [10, 20, 30, 40]
     out: 100
     lines: { '2': 1, 4: 1}
-    statements: {'1': 1, '2': 1 }
-    functions: {'1': 1}
-    branches: { '1': [0], '2': [0], '3': [0], '4': [0] }
+    statements: {'0': 1, '1': 1 }
+    functions: {'0': 1}
+    branches: { '0': [0], '1': [0], '2': [0], '3': [0] }
 
   - name: 2 of 4 specified
     args: [3, 4 ]
     out: 14
     lines: { '2': 1, 4: 1}
-    statements: {'1': 1, '2': 1 }
-    functions: {'1': 1}
-    branches: { '1': [0], '2': [0], '3': [1], '4': [1] }
+    statements: {'0': 1, '1': 1 }
+    functions: {'0': 1}
+    branches: { '0': [0], '1': [0], '2': [1], '3': [1] }
 
   - name: nothing specified
     args: []
     out: 10
     lines: { '2': 1, 4: 1}
-    statements: {'1': 1, '2': 1 }
-    functions: {'1': 1}
-    branches: { '1': [1], '2': [1], '3': [1], '4': [1] }
+    statements: {'0': 1, '1': 1 }
+    functions: {'0': 1}
+    branches: { '0': [1], '1': [1], '2': [1], '3': [1] }