--- a/test/specs/do-while.yaml
+++ b/test/specs/do-while.yaml
@@ -9,13 +9,13 @@ tests:
     args: [10]
     out: 10
     lines: {'1': 1, '2': 10, '3': 1}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 10, '5': 1}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 10, '4': 1}
 
   - name: single entry into while
     args: [-1]
     out: 1
     lines: {'1': 1, '2': 1, '3': 1}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 1}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 1}
 
 ---
 name: block do-while on separate line
@@ -30,13 +30,13 @@ tests:
     args: [10]
     out: 10
     lines: {'1': 1, '2': 1, '3': 10, '5': 1}
-    statements: {'1': 1, '2':1, '3': 1, '4': 10, '5': 1}
+    statements: {'0': 1, '1':1, '2': 1, '3': 10, '4': 1}
 
   - name: single entry into while
     args: [-1]
     out: 1
     lines: {'1': 1, '2': 1, '3': 1, '5': 1}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 1}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 1}
 
 ---
 name: ignore inside do-while
@@ -50,4 +50,4 @@ tests:
   - args: []
     out: 10
     lines: {'1': 1, '5': 1}
-    statements: {'1': 1, '2': 1 }
+    statements: {'0': 1, '1': 1 }