--- a/test/specs/for-of.yaml
+++ b/test/specs/for-of.yaml
@@ -12,8 +12,8 @@ tests:
   - args: []
     out: 3
     lines: {'1': 1, '3': 1, '4': 1, '5': 2}
-    functions: {'1': 1}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 2}
+    functions: {'0': 1}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 2}
 ---
 
 name: for-of with loop initializer
@@ -28,5 +28,5 @@ tests:
   - args: []
     out: 3
     lines: {'1': 1, '2': 1, '3': 1, '4': 2}
-    functions: {'1': 1}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 2}
+    functions: {'0': 1}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 2}