--- a/test/specs/functions.yaml
+++ b/test/specs/functions.yaml
@@ -11,17 +11,17 @@ tests:
     args: [2]
     out: 42
     lines: {'1': 1, '3': 1, '5': 1}
-    branches: {'1': [1, 0]}
-    functions: {'1': 1}
-    statements: {'1': 1, '2': 1, '3': 1 }
+    branches: {'0': [1, 0]}
+    functions: {'0': 1}
+    statements: {'0': 1, '1': 1, '2': 1 }
 
   - name: does not cover function
     args: [10]
     out: 15
     lines: {'1': 1, '3': 0, '5': 1}
-    branches: {'1': [0, 1]}
-    functions: {'1': 0}
-    statements: {'1': 1, '2': 0, '3': 1}
+    branches: {'0': [0, 1]}
+    functions: {'0': 0}
+    statements: {'0': 1, '1': 0, '2': 1 }
 
 ---
 name: anonymous function
@@ -33,17 +33,17 @@ tests:
     args: [2]
     out: 42
     lines: {'1': 1, '2': 1}
-    branches: {'1': [1, 0]}
-    functions: {'1': 1}
-    statements: {'1': 1, '2': 1, '3': 1}
+    branches: {'0': [1, 0]}
+    functions: {'0': 1}
+    statements: {'0': 1, '1': 1, '2': 1 }
 
   - name: does not cover function
     args: [10]
     out: 15
     lines: {'1': 1, '2': 1}
-    branches: {'1': [0, 1]}
-    functions: {'1': 0}
-    statements: {'1': 1, '2': 1, '3': 0}
+    branches: {'0': [0, 1]}
+    functions: {'0': 0}
+    statements: {'0': 1, '1': 1, '2': 0}
 
 ---
 name: anonymous function newline
@@ -59,17 +59,17 @@ tests:
     args: [2]
     out: 42
     lines: {'1': 1, '2': 1, '4': 1}
-    branches: {'1': [1, 0]}
-    functions: {'1': 1}
-    statements: {'1': 1, '2': 1, '3': 1}
+    branches: {'0': [1, 0]}
+    functions: {'0': 1}
+    statements: {'0': 1, '1': 1, '2': 1}
 
   - name: does not cover function
     args: [10]
     out: 15
     lines: {'1': 1, '2': 1, '4': 0}
-    branches: {'1': [0, 1]}
-    functions: {'1': 0}
-    statements: {'1': 1, '2': 1, '3': 0}
+    branches: {'0': [0, 1]}
+    functions: {'0': 0}
+    statements: {'0': 1, '1': 1, '2': 0}
 
 ---
 name: function decl in unreachable place
@@ -84,14 +84,14 @@ tests:
     args: [1]
     out: 2
     lines: { '2': 0, '3': 0, '5': 1}
-    branches: {'1': [1, 0]}
-    functions: {'1': 0, '2': 0}
-    statements: {'1': 0, '2': 0, '3': 1}
+    branches: {'0': [1, 0]}
+    functions: {'0': 0, '1': 0}
+    statements: {'0': 0, '1': 0, '2': 1}
 
   - name: covers declaration and function
     args: [10]
     out: 20
     lines: { '2': 1, '3': 1, '5': 1}
-    branches: {'1': [0, 1]}
-    functions: {'1': 1, '2': 1}
-    statements: {'1': 1, '2': 1, '3': 1 }
+    branches: {'0': [0, 1]}
+    functions: {'0': 1, '1': 1}
+    statements: {'0': 1, '1': 1, '2': 1 }