--- a/test/specs/statement-hints.yaml
+++ b/test/specs/statement-hints.yaml
@@ -10,7 +10,7 @@ tests:
     out: 10
     lines: {'3': 1}
     branches: {}
-    statements: {'1': 1}
+    statements: {'0': 1}
 
 ---
 name: ignore before function declaration
@@ -23,7 +23,7 @@ tests:
     out: 10
     lines: { '3': 1 }
     functions: {}
-    statements: { '1': 1 }
+    statements: { '0': 1 }
 
 ---
 name: ignore before func expressions
@@ -64,8 +64,8 @@ tests:
   - args: ['2']
     out: 1
     lines: {'1': 1, '4': 1}
-    branches: {'1': [1]}
-    statements: {'1': 1, '2': 1}
+    branches: {'0': [1]}
+    statements: {'0': 1, '1': 1}
 
 ---
 name: ignore before ternary statement
@@ -87,8 +87,8 @@ tests:
   - args: [2]
     out: 0
     lines: {'1': 1}
-    branches: {'1': [1]}
-    statements: {'1': 1}
+    branches: {'0': [1]}
+    statements: {'0': 1}
 
 ---
 name: ignore in logical expression
@@ -102,8 +102,8 @@ tests:
   - args: [1]
     out: 11
     lines: {'1': 1, '2': 1, '4': 0}
-    branches: {'1': [1, 0], '2': [1]}
-    statements: {'1': 1, '2': 1, '3': 0}
+    branches: {'0': [1, 0], '1': [1]}
+    statements: {'0': 1, '1': 1, '2': 0}
 
 ---
 name: ignore in complex logical expression
@@ -117,8 +117,8 @@ tests:
   - args: [1]
     out: 11
     lines: {'1': 1, '2': 1, '4': 0}
-    branches: {'1': [1, 0], '2': [1]}
-    statements: {'1': 1, '2': 1, '3': 0}
+    branches: {'0': [1, 0], '1': [1]}
+    statements: {'0': 1, '1': 1, '2': 0}
 
 ---
 name: ignore in logical expression with implied operator precedence
@@ -132,8 +132,8 @@ tests:
   - args: [1, 1]
     out: 11
     lines: {'1': 1, '2': 1, '4': 0}
-    branches: {'1': [1, 0], '2': [1]}
-    statements: {'1': 1, '2': 1, '3': 0}
+    branches: {'0': [1, 0], '1': [1]}
+    statements: {'0': 1, '1': 1, '2': 0}
 
 ---
 name: ignore before logical expression
@@ -147,5 +147,5 @@ tests:
   - args: [1, 1]
     out: 11
     lines: {'1': 1, '2': 1, '4': 0}
-    branches: {'1': [1, 0] }
-    statements: {'1': 1, '2': 1, '3': 0}
+    branches: {'0': [1, 0] }
+    statements: {'0': 1, '1': 1, '2': 0}