--- a/test/specs/statement.yaml
+++ b/test/specs/statement.yaml
@@ -9,7 +9,7 @@ tests:
     out: 10
     lines: {'2': 1}
     branches: {}
-    statements: {'1': 1}
+    statements: {'0': 1}
 ---
 name: simple statement
 code: |
@@ -20,15 +20,15 @@ tests:
     args: [10]
     out: 10
     lines: {'1': 1, '2': 1}
-    branches: {'1': [1, 0]}
-    statements: {'1': 1, '2': 1}
+    branches: {'0': [1, 0]}
+    statements: {'0': 1, '1': 1}
 
   - name: covers line and other branch
     args: [1]
     out: undef
     lines: {'1': 1, '2': 1}
-    branches: {'1': [0, 1]}
-    statements: {'1': 1, '2': 1}
+    branches: {'0': [0, 1]}
+    statements: {'0': 1, '1': 1}
 ---
 name: shebang code
 code: |
@@ -40,15 +40,15 @@ tests:
     args: [10]
     out: 10
     lines: {'2': 1, '3': 1}
-    branches: {'1': [1, 0]}
-    statements: {'1': 1, '2': 1}
+    branches: {'0': [1, 0]}
+    statements: {'0': 1, '1': 1}
 
   - name: covers line and other branch
     args: [1]
     out: undef
     lines: {'2': 1, '3': 1}
-    branches: {'1': [0, 1]}
-    statements: {'1': 1, '2': 1}
+    branches: {'0': [0, 1]}
+    statements: {'0': 1, '1': 1}
 ---
 name: mainline return
 instrumentOpts:
@@ -60,4 +60,4 @@ tests:
     args: []
     out: 10
     lines: {'1': 1 }
-    statements: { '1': 1 }
+    statements: { '0': 1 }