--- a/test/specs/strict.yaml
+++ b/test/specs/strict.yaml
@@ -16,9 +16,9 @@ tests:
   - name: covers one statement less
     args: []
     out: pass
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 0, '6': 1}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 0, '5': 1}
     lines: {'1': 1, '3': 1, '4': 1, '5': 1, '6': 0, '8': 1}
-    functions: {'1': 1}
+    functions: {'0': 1}
 
 ---
 name: function decl using strict
@@ -39,9 +39,9 @@ tests:
   - name: covers one statement less
     args: []
     out: pass
-    statements: {'1': 1, '2': 1, '3': 1, '4': 0, '5': 1, '6': 1 }
+    statements: {'0': 1, '1': 1, '2': 1, '3': 0, '4': 1, '5': 1 }
     lines: { '3': 1, '4': 1, '5': 1, '6': 0, '8': 1, '11': 1}
-    functions: {'1': 1}
+    functions: {'0': 1}
 
 ---
 name: function decl that looks like strict but is not
@@ -63,9 +63,9 @@ tests:
   - name: covers all statements as usual
     args: []
     out: fail
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 1, '6': 1, '7': 0, '8': 1 }
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 1, '5': 1, '6': 0, '7': 1 }
     lines: { '2': 1, '3': 1, '4': 1, '5': 1, '6': 1, '7': 1, '9': 0, '12': 1}
-    functions: {'1': 1}
+    functions: {'0': 1}
 
 ---
 name: file-level strict declaration
@@ -82,5 +82,5 @@ code: |
 tests:
   - args: []
     out: fail
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 0}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 0}
     lines: {'2': 1, '3': 1, '4': 1, '5': 1, '7': 0}