--- a/test/specs/switch.yaml
+++ b/test/specs/switch.yaml
@@ -8,7 +8,7 @@ tests:
     args: ['1']
     out: unknown
     lines: {'1': 1, '2': 1}
-    statements: {'1': 1, '2': 1}
+    statements: {'0': 1, '1': 1}
 
 ---
 name: 2 cases no default
@@ -23,22 +23,22 @@ tests:
     args: ['1']
     out: one
     lines: {'1': 1, '2': 1, '3': 1, '4': 0}
-    branches: {'1': [1, 0]}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 0, '6': 0}
+    branches: {'0': [1, 0]}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 0, '5': 0}
 
   - name: second case
     args: ['2']
     out: two
     lines: {'1': 1, '2': 1, '3': 0, '4': 1}
-    branches: {'1': [0, 1]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 1, '6': 1}
+    branches: {'0': [0, 1]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 0, '4': 1, '5': 1}
 
   - name: no match
     args: ['3']
     out: unknown
     lines: {'1': 1, '2': 1, '3': 0, '4': 0}
-    branches: {'1': [0, 0]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 0, '6': 0}
+    branches: {'0': [0, 0]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 0, '4': 0, '5': 0}
 
 ---
 name: 2 cases with default
@@ -54,22 +54,22 @@ tests:
     args: ['1']
     out: one
     lines: {'1': 1, '2': 1, '3': 1, '4': 0, '5': 0}
-    branches: {'1': [1, 0, 0]}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 0, '6': 0, '7': 0}
+    branches: {'0': [1, 0, 0]}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 0, '5': 0, '6': 0}
 
   - name: second case
     args: ['2']
     out: two
     lines: {'1': 1, '2': 1, '3': 0, '4': 1, '5': 0}
-    branches: {'1': [0, 1, 0]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 1, '6': 1, '7': 0}
+    branches: {'0': [0, 1, 0]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 0, '4': 1, '5': 1, '6': 0}
 
   - name: default case
     args: ['4']
     out: three
     lines: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 1}
-    branches: {'1': [0, 0, 1]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 0, '6': 0, '7': 1}
+    branches: {'0': [0, 0, 1]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 0, '4': 0, '5': 0, '6': 1}
 
 ---
 name: one line layout
@@ -81,22 +81,22 @@ tests:
     args: ['1']
     out: one
     lines: {'1': 1, '2': 1}
-    branches: {'1': [1, 0, 0]}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 0, '6': 0, '7': 0}
+    branches: {'0': [1, 0, 0]}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 0, '5': 0, '6': 0}
 
   - name: second case
     args: ['2']
     out: two
     lines: {'1': 1, '2': 1}
-    branches: {'1': [0, 1, 0]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 1, '6': 1, '7': 0}
+    branches: {'0': [0, 1, 0]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 0, '4': 1, '5': 1, '6': 0}
 
   - name: default case
     args: ['4']
     out: three
     lines: {'1': 1, '2': 1}
-    branches: {'1': [0, 0, 1]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 0, '6': 0, '7': 1}
+    branches: {'0': [0, 0, 1]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 0, '4': 0, '5': 0, '6': 1}
 
 ---
 name: 2 cases with default and fallthru
@@ -112,22 +112,22 @@ tests:
     args: ['1']
     out: onetwothree
     lines: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 1}
-    branches: {'1': [1, 1, 1]}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 1}
+    branches: {'0': [1, 1, 1]}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 1}
 
   - name: second case
     args: ['2']
     out: twothree
     lines: {'1': 1, '2': 1, '3': 0, '4': 1, '5': 1}
-    branches: {'1': [0, 1, 1]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 1, '5': 1}
+    branches: {'0': [0, 1, 1]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 1, '4': 1}
 
   - name: default case
     args: ['4']
     out: three
     lines: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 1}
-    branches: {'1': [0, 0, 1]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 1}
+    branches: {'0': [0, 0, 1]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 0, '4': 1}
 
 ---
 name: one-line layout with fallthru
@@ -139,19 +139,19 @@ tests:
     args: ['1']
     out: onetwothree
     lines: {'1': 1, '2': 1}
-    branches: {'1': [1, 1, 1]}
-    statements: {'1': 1, '2': 1, '3': 1, '4': 1, '5': 1}
+    branches: {'0': [1, 1, 1]}
+    statements: {'0': 1, '1': 1, '2': 1, '3': 1, '4': 1}
 
   - name: second case
     args: ['2']
     out: twothree
     lines: {'1': 1, '2': 1}
-    branches: {'1': [0, 1, 1]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 1, '5': 1}
+    branches: {'0': [0, 1, 1]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 1, '4': 1}
 
   - name: default case
     args: ['4']
     out: three
     lines: {'1': 1, '2': 1}
-    branches: {'1': [0, 0, 1]}
-    statements: {'1': 1, '2': 1, '3': 0, '4': 0, '5': 1}
+    branches: {'0': [0, 0, 1]}
+    statements: {'0': 1, '1': 1, '2': 0, '3': 0, '4': 1}