Closes #21. Makes instrumented code run much faster by making V8 use “fast” mode when accessing stats objects, which wasn't triggered before because of non-0-based indices.

Here's are the timings from [mapbox-gl-js render suite](https://github.com/mapbox/mapbox-gl-js):

- **no coverage**: 1m 6s
- **nyc current**: 4m 37s (or 3m 31s overhead)
- **nyc updated**: 1m 20s (or 14s overhead)

So, basically the performance overhead is 15 times smaller, down from 320% to 20%.

- [x] update tests

cc @bcoe @mramato