@mourner 2016-09-07
@bcoe @gotwarlost anything I could help with to get this merged? A 15x speedup seems like something that's worth prioritizing. :)