@bcoe 2016-09-08
@dmitriiabramov @mourner I see no reason not to land this; but since I haven't heard back from @gotwarlost, I will be testing quite throughly; I'll start by publishing this to a `next` tag.