@bcoe 2016-09-08
@mourner I've published these changes to the `next` tag on npm, if you'd like to help test a bit:

`npm cache clear; npm i istanbul-lib-instrument@next`