@cpojer 2016-09-08
Oh this is awesome! Is anyone interested in trying this out on https://github.com/facebook/jest itself? :) (run jest with --coverage).