@mramato 2016-09-08
Thanks @mourner, I would love to try this out to confirm my issue in https://github.com/gotwarlost/istanbul/issues/556 is gone.  Unfortunately I currently rely on https://github.com/karma-runner/karma-coverage which hasn't updated to this library yet (but I'm hoping it happens soon or I can open a PR to do it for them).