@dmitriiabramov 2016-09-08
@cpojer i tested it on jest itself, but we don't really instrument much code, so there isn't much we can optimize :)
although it seems to run 1 second faster, 28s on average instead of 29s