@mourner 2016-09-08
@bcoe thanks! The new version works perfectly in our project:

- [no coverage](https://circleci.com/gh/mapbox/mapbox-gl-js/4615) — `test.sh` takes 4m53s
- [nyc latest](https://circleci.com/gh/mapbox/mapbox-gl-js/4906) — `test.sh` takes 11m22s
- [nyc new](https://circleci.com/gh/mapbox/mapbox-gl-js/4908) — `test.sh` takes 5m56s