@gotwarlost 2016-10-02
So sorry about my lack of response (this is getting old :( )

I see no reason off the top of my head why this should not be merged. 👍